Add optimizations for getting and invoking dotted properties.
This is a more general solution that covers getters and generalized dotted invocations. Admittedly, some of the optimization code in PXByteCodeBuilder (particularly addGetPropertyInstructionWithName) is a little ugly. We may want to consider a separate optimization pass that captures all of this logic elsewhere, thus making this step optional as well.